improvement
Use async and await JavaScript features to consume the API and organize function in different files
